consolidate move graphical object action code
